refactor: Output 메서드 축소, io 디렉터리로 이동
