+++ spring-basic/src/main/java/org/prgrms/springbasic/service/console/customer/RegisterBlackService.java
@@ -2,13 +2,14 @@ package org.prgrms.springbasic.service.console.customer;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.prgrms.springbasic.domain.console.Console;
+import org.prgrms.springbasic.utils.io.console.Console;
 import org.prgrms.springbasic.domain.customer.Customer;
 import org.prgrms.springbasic.repository.customer.CustomerRepository;
 import org.prgrms.springbasic.service.console.ConsoleService;
 import org.springframework.stereotype.Service;
 
-import java.util.UUID;
+import static java.util.UUID.randomUUID;
+import static org.prgrms.springbasic.utils.enumm.ConsoleMessage.REGISTER_BLACK;
 
 
 @Service("register")
@@ -21,9 +22,9 @@ public class RegisterBlackService implements ConsoleService {
 
     @Override
     public void execute() {
-        console.printNameInput();
+        console.printToConsole(REGISTER_BLACK.getMessage());
         var name = console.takeInput();
-        var newBlackCustomer =  Customer.blackCustomer(UUID.randomUUID(), name);
+        var newBlackCustomer =  Customer.blackCustomer(randomUUID(), name);
         repository.save(newBlackCustomer);
     }
