+++ spring-basic/src/main/java/org/prgrms/springbasic/utils/io/console/Input.java
@@ -1,4 +1,4 @@
-package org.prgrms.springbasic.domain.console;
+package org.prgrms.springbasic.utils.io.console;
 
 public interface Input {
     String takeInput();
