+++ spring-basic/src/main/java/org/prgrms/springbasic/utils/io/console/Output.java
@@ -1,0 +1,5 @@
+package org.prgrms.springbasic.utils.io.console;
+
+public interface Output {
+    void printToConsole(Object object);
+}
