+++ spring-basic/src/test/java/org/prgrms/springbasic/service/console/customer/ListBlackServiceTest.java
@@ -2,12 +2,11 @@ package org.prgrms.springbasic.service.console.customer;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.prgrms.springbasic.domain.console.Console;
+import org.prgrms.springbasic.utils.io.console.Console;
 import org.prgrms.springbasic.repository.customer.BlackCustomerRepository;
 import org.prgrms.springbasic.repository.customer.CustomerRepository;
 import org.prgrms.springbasic.service.console.ConsoleService;
 
-import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
 import java.util.List;
